requirements.txt
backcall==0.1.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
dynalist==0.2.6
dynapi==0.0.2
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
PyYAML==5.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
smmap==3.0.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
